requirements.txt
Flask==2.2.5
azure-ai-inference==1.0.0b9
azure-core==1.32.0
